+++ tests/scripts/script_all_errors.py
@@ -19,8 +19,6 @@ if __name__ == "__main__":
         weechat.hook_completion_get_string("0x123abc", "base_command")
         weechat.hook_completion_list_add("0x123abc", "word", 0,
                                          weechat.WEECHAT_LIST_POS_SORT)
-        color_code = weechat.info_get("irc_nick_color", "nick")
-        color_name = weechat.info_get("irc_nick_color_name", "nick")
         weechat.hook_modifier("irc_in_privmsg", "callback", "")
         weechat.hook_signal("*,irc_out_privmsg", "callback", "")
         weechat.hook_signal("*,irc_outtags_privmsg", "callback", "")