+++ tests/test_main.py
@@ -102,7 +102,7 @@ def test_main_dir() -> None:
         with mock.patch.object(sys, "argv", args):
             weechat_script_lint.main()
     assert exc.type == SystemExit
-    assert exc.value.code == 26
+    assert exc.value.code == 22
 
     # check directory with scripts, display scripts by score
     args = [