+++ tests/test_script.py
@@ -31,14 +31,12 @@ ALL_ERRORS = [
     ("error", 17, "missing_infolist_free"),
     ("error", 18, "python2_bin"),
     ("error", 1, "mixed_tabs_spaces"),
-    ("warning", 27, "sys_exit"),
+    ("warning", 25, "sys_exit"),
     ("warning", 19, "deprecated_hook_completion_get_string"),
     ("warning", 20, "deprecated_hook_completion_list_add"),
-    ("warning", 22, "deprecated_irc_nick_color"),
-    ("warning", 23, "deprecated_irc_nick_color_name"),
-    ("warning", 24, "modifier_irc_in"),
-    ("warning", 25, "signal_irc_out"),
-    ("warning", 26, "signal_irc_outtags"),
+    ("warning", 22, "modifier_irc_in"),
+    ("warning", 23, "signal_irc_out"),
+    ("warning", 24, "signal_irc_outtags"),
     ("info", 1, "unneeded_shebang"),
     ("info", 13, "url_weechat"),
 ]
@@ -79,12 +77,12 @@ def test_script_all_errors() -> None:
     assert script.script
     script.check()
     assert str(script)
-    assert len(str(script).split("\n")) == 14
+    assert len(str(script).split("\n")) == 12
     print(script.count)
-    assert script.count == {"error": 4, "warning": 8, "info": 2}
+    assert script.count == {"error": 4, "warning": 6, "info": 2}
     errors = [(msg.level, msg.line, msg.msg_name) for msg in script.messages]
     assert errors == ALL_ERRORS
-    assert len(script.get_report(False).split("\n")) == 14
+    assert len(script.get_report(False).split("\n")) == 12
     assert script.get_report(True) == "script_all_errors.py"
 
     # ignore 2 messages: "missing_email" and "sys_exit"
@@ -99,9 +97,9 @@ def test_script_all_errors() -> None:
     assert script.script
     script.check()
     assert str(script)
-    assert len(str(script).split("\n")) == 12
-    assert script.count == {"error": 3, "warning": 7, "info": 2}
-    assert len(script.get_report(False).split("\n")) == 12
+    assert len(str(script).split("\n")) == 10
+    assert script.count == {"error": 3, "warning": 5, "info": 2}
+    assert len(script.get_report(False).split("\n")) == 10
     assert script.get_report(True) == "script_all_errors.py"